@@ -1,9 +1,9 @@
 module github.com/awslabs/cdk8s-go/cdk8splus17
 
-go 1.15
+go 1.16
 
 require (
-	github.com/aws/jsii-runtime-go v1.20.1
-	github.com/awslabs/cdk8s-go/cdk8s v1.0.0-beta.9
-	github.com/aws/constructs-go/constructs/v3 v3.3.5
+	github.com/aws/jsii-runtime-go v1.23.0
+	github.com/awslabs/cdk8s-go/cdk8s v1.0.0-beta.10
+	github.com/aws/constructs-go/constructs/v3 v3.3.48
 )
