@@ -4,210 +4,129 @@ package cdk8s
 import (
 	_jsii_ "github.com/aws/jsii-runtime-go"
 	_init_ "github.com/awslabs/cdk8s-go/cdk8s/jsii"
-	"reflect"
+
 	"github.com/aws/constructs-go/constructs/v3"
 )
 
-// Class interface
-type ApiObjectIface interface {
-	constructs.IConstructIface
-	GetApiGroup() string
-	GetApiVersion() string
-	GetChart() ChartIface
-	GetKind() string
-	GetMetadata() ApiObjectMetadataDefinitionIface
-	GetName() string
-	OnPrepare()
-	OnSynthesize(session constructs.ISynthesisSessionIface)
-	OnValidate() []string
-	ToString() string
-	AddDependency(dependencies constructs.IConstructIface)
-	AddJsonPatch(ops JsonPatchIface)
+// Experimental.
+type ApiObject interface {
+	constructs.Construct
+	ApiGroup() string
+	ApiVersion() string
+	Chart() Chart
+	Kind() string
+	Metadata() ApiObjectMetadataDefinition
+	Name() string
+	AddDependency(dependencies constructs.IConstruct)
+	AddJsonPatch(ops JsonPatch)
 	ToJson() interface{}
 }
 
-// Experimental.
-// Struct proxy
-type ApiObject struct {
-	// The group portion of the API version (e.g. `authorization.k8s.io`).
-	// Experimental.
-	ApiGroup string `json:"apiGroup"`
-	// The object's API version (e.g. `authorization.k8s.io/v1`).
-	// Experimental.
-	ApiVersion string `json:"apiVersion"`
-	// The chart in which this object is defined.
-	// Experimental.
-	Chart ChartIface `json:"chart"`
-	// The object kind.
-	// Experimental.
-	Kind string `json:"kind"`
-	// Metadata associated with this API object.
-	// Experimental.
-	Metadata ApiObjectMetadataDefinitionIface `json:"metadata"`
-	// The name of the API object.
-	//
-	// If a name is specified in `metadata.name` this will be the name returned.
-	// Otherwise, a name will be generated by calling
-	// `Chart.of(this).generatedObjectName(this)`, which by default uses the
-	// construct path to generate a DNS-compatible name for the resource.
-	// Experimental.
-	Name string `json:"name"`
+// The jsii proxy struct for ApiObject
+type apiObject struct {
+	constructs.Construct // extends constructs.Construct
 }
 
-func (a *ApiObject) GetApiGroup() string {
+func (a *apiObject) ApiGroup() string {
 	var returns string
 	_jsii_.Get(
 		a,
 		"apiGroup",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (a *ApiObject) GetApiVersion() string {
+func (a *apiObject) ApiVersion() string {
 	var returns string
 	_jsii_.Get(
 		a,
 		"apiVersion",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (a *ApiObject) GetChart() ChartIface {
-	var returns ChartIface
+func (a *apiObject) Chart() Chart {
+	var returns Chart
 	_jsii_.Get(
 		a,
 		"chart",
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*ChartIface)(nil)).Elem(): reflect.TypeOf((*Chart)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (a *ApiObject) GetKind() string {
+func (a *apiObject) Kind() string {
 	var returns string
 	_jsii_.Get(
 		a,
 		"kind",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (a *ApiObject) GetMetadata() ApiObjectMetadataDefinitionIface {
-	var returns ApiObjectMetadataDefinitionIface
+func (a *apiObject) Metadata() ApiObjectMetadataDefinition {
+	var returns ApiObjectMetadataDefinition
 	_jsii_.Get(
 		a,
 		"metadata",
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*ApiObjectMetadataDefinitionIface)(nil)).Elem(): reflect.TypeOf((*ApiObjectMetadataDefinition)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (a *ApiObject) GetName() string {
+func (a *apiObject) Name() string {
 	var returns string
 	_jsii_.Get(
 		a,
 		"name",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
 
 // Defines an API object.
-func NewApiObject(scope constructs.ConstructIface, id string, props ApiObjectPropsIface) ApiObjectIface {
+func NewApiObject(scope constructs.Construct, id string, props ApiObjectProps) ApiObject {
 	_init_.Initialize()
-	self := ApiObject{}
+	a := apiObject{}
+
 	_jsii_.Create(
 		"cdk8s.ApiObject",
 		[]interface{}{scope, id, props},
 		[]_jsii_.FQN{},
 		[]_jsii_.Override{},
-		&self,
+		&a,
 	)
-	return &self
+	return &a
 }
 
-func ApiObject_Of(c constructs.IConstructIface) ApiObjectIface {
+// Returns the `ApiObject` named `Resource` which is a child of the given construct.
+//
+// If `c` is an `ApiObject`, it is returned directly. Throws an
+// exception if the construct does not have a child named `Default` _or_ if
+// this child is not an `ApiObject`.
+// Experimental.
+func ApiObject_Of(c constructs.IConstruct) ApiObject {
 	_init_.Initialize()
-	var returns ApiObjectIface
-	_jsii_.InvokeStatic(
+	var returns ApiObject
+	_jsii_.StaticInvoke(
 		"cdk8s.ApiObject",
 		"of",
 		[]interface{}{c},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*ApiObjectIface)(nil)).Elem(): reflect.TypeOf((*ApiObject)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (a *ApiObject) OnPrepare() {
-	var returns interface{}
-	_jsii_.Invoke(
-		a,
-		"onPrepare",
-		[]interface{}{},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (a *ApiObject) OnSynthesize(session constructs.ISynthesisSessionIface) {
-	var returns interface{}
-	_jsii_.Invoke(
-		a,
-		"onSynthesize",
-		[]interface{}{session},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (a *ApiObject) OnValidate() []string {
-	var returns []string
-	_jsii_.Invoke(
-		a,
-		"onValidate",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (a *ApiObject) ToString() string {
-	var returns string
-	_jsii_.Invoke(
-		a,
-		"toString",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (a *ApiObject) AddDependency(dependencies constructs.IConstructIface) {
+// Create a dependency between this ApiObject and other constructs.
+//
+// These can be other ApiObjects, Charts, or custom.
+// Experimental.
+func (a *apiObject) AddDependency(dependencies constructs.IConstruct) {
 	var returns interface{}
 	_jsii_.Invoke(
 		a,
@@ -215,11 +134,15 @@ func (a *ApiObject) AddDependency(dependencies constructs.IConstructIface) {
 		[]interface{}{dependencies},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
-func (a *ApiObject) AddJsonPatch(ops JsonPatchIface) {
+// Applies a set of RFC-6902 JSON-Patch operations to the manifest synthesized for this API object.
+//
+// TODO: EXAMPLE
+//
+// Experimental.
+func (a *apiObject) AddJsonPatch(ops JsonPatch) {
 	var returns interface{}
 	_jsii_.Invoke(
 		a,
@@ -227,11 +150,15 @@ func (a *ApiObject) AddJsonPatch(ops JsonPatchIface) {
 		[]interface{}{ops},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
-func (a *ApiObject) ToJson() interface{} {
+// Renders the object to Kubernetes JSON.
+//
+// To disable sorting of dictionary keys in output object set the
+// `CDK8S_DISABLE_SORT` environment variable to any non-empty value.
+// Experimental.
+func (a *apiObject) ToJson() interface{} {
 	var returns interface{}
 	_jsii_.Invoke(
 		a,
@@ -239,23 +166,12 @@ func (a *ApiObject) ToJson() interface{} {
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-// ApiObjectMetadataIface is the public interface for the custom type ApiObjectMetadata
-// Experimental.
-type ApiObjectMetadataIface interface {
-	GetAnnotations() map[string]string
-	GetLabels() map[string]string
-	GetName() string
-	GetNamespace() string
-}
-
 // Metadata associated with this object.
 // Experimental.
-// Struct proxy
 type ApiObjectMetadata struct {
 	// Annotations is an unstructured key value map stored with a resource that may be set by external tools to store and retrieve arbitrary metadata.
 	//
@@ -296,59 +212,11 @@ type ApiObjectMetadata struct {
 	Namespace string `json:"namespace"`
 }
 
-func (a *ApiObjectMetadata) GetAnnotations() map[string]string {
-	var returns map[string]string
-	_jsii_.Get(
-		a,
-		"annotations",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (a *ApiObjectMetadata) GetLabels() map[string]string {
-	var returns map[string]string
-	_jsii_.Get(
-		a,
-		"labels",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (a *ApiObjectMetadata) GetName() string {
-	var returns string
-	_jsii_.Get(
-		a,
-		"name",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (a *ApiObjectMetadata) GetNamespace() string {
-	var returns string
-	_jsii_.Get(
-		a,
-		"namespace",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-
-// Class interface
-type ApiObjectMetadataDefinitionIface interface {
-	GetName() string
-	GetNamespace() string
+// Object metadata.
+// Experimental.
+type ApiObjectMetadataDefinition interface {
+	Name() string
+	Namespace() string
 	Add(key string, value interface{})
 	AddAnnotation(key string, value string)
 	AddLabel(key string, value string)
@@ -356,60 +224,49 @@ type ApiObjectMetadataDefinitionIface interface {
 	ToJson() interface{}
 }
 
-// Object metadata.
-// Experimental.
-// Struct proxy
-type ApiObjectMetadataDefinition struct {
-	// The name of the API object.
-	//
-	// If a name is specified in `metadata.name` this will be the name returned.
-	// Otherwise, a name will be generated by calling
-	// `Chart.of(this).generatedObjectName(this)`, which by default uses the
-	// construct path to generate a DNS-compatible name for the resource.
-	// Experimental.
-	Name string `json:"name"`
-	// The object's namespace.
-	// Experimental.
-	Namespace string `json:"namespace"`
+// The jsii proxy struct for ApiObjectMetadataDefinition
+type apiObjectMetadataDefinition struct {
+	_ byte // padding
 }
 
-func (a *ApiObjectMetadataDefinition) GetName() string {
+func (a *apiObjectMetadataDefinition) Name() string {
 	var returns string
 	_jsii_.Get(
 		a,
 		"name",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (a *ApiObjectMetadataDefinition) GetNamespace() string {
+func (a *apiObjectMetadataDefinition) Namespace() string {
 	var returns string
 	_jsii_.Get(
 		a,
 		"namespace",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
 
-func NewApiObjectMetadataDefinition(options ApiObjectMetadataIface) ApiObjectMetadataDefinitionIface {
+func NewApiObjectMetadataDefinition(options ApiObjectMetadata) ApiObjectMetadataDefinition {
 	_init_.Initialize()
-	self := ApiObjectMetadataDefinition{}
+	a := apiObjectMetadataDefinition{}
+
 	_jsii_.Create(
 		"cdk8s.ApiObjectMetadataDefinition",
 		[]interface{}{options},
 		[]_jsii_.FQN{},
 		[]_jsii_.Override{},
-		&self,
+		&a,
 	)
-	return &self
+	return &a
 }
 
-func (a *ApiObjectMetadataDefinition) Add(key string, value interface{}) {
+// Adds an arbitrary key/value to the object metadata.
+// Experimental.
+func (a *apiObjectMetadataDefinition) Add(key string, value interface{}) {
 	var returns interface{}
 	_jsii_.Invoke(
 		a,
@@ -417,11 +274,12 @@ func (a *ApiObjectMetadataDefinition) Add(key string, value interface{}) {
 		[]interface{}{key, value},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
-func (a *ApiObjectMetadataDefinition) AddAnnotation(key string, value string) {
+// Add an annotation.
+// Experimental.
+func (a *apiObjectMetadataDefinition) AddAnnotation(key string, value string) {
 	var returns interface{}
 	_jsii_.Invoke(
 		a,
@@ -429,11 +287,12 @@ func (a *ApiObjectMetadataDefinition) AddAnnotation(key string, value string) {
 		[]interface{}{key, value},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
-func (a *ApiObjectMetadataDefinition) AddLabel(key string, value string) {
+// Add a label.
+// Experimental.
+func (a *apiObjectMetadataDefinition) AddLabel(key string, value string) {
 	var returns interface{}
 	_jsii_.Invoke(
 		a,
@@ -441,11 +300,12 @@ func (a *ApiObjectMetadataDefinition) AddLabel(key string, value string) {
 		[]interface{}{key, value},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
-func (a *ApiObjectMetadataDefinition) GetLabel(key string) string {
+// Returns: a value of a label or undefined
+// Experimental.
+func (a *apiObjectMetadataDefinition) GetLabel(key string) string {
 	var returns string
 	_jsii_.Invoke(
 		a,
@@ -453,12 +313,13 @@ func (a *ApiObjectMetadataDefinition) GetLabel(key string) string {
 		[]interface{}{key},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (a *ApiObjectMetadataDefinition) ToJson() interface{} {
+// Synthesizes a k8s ObjectMeta for this metadata set.
+// Experimental.
+func (a *apiObjectMetadataDefinition) ToJson() interface{} {
 	var returns interface{}
 	_jsii_.Invoke(
 		a,
@@ -466,22 +327,12 @@ func (a *ApiObjectMetadataDefinition) ToJson() interface{} {
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-// ApiObjectPropsIface is the public interface for the custom type ApiObjectProps
-// Experimental.
-type ApiObjectPropsIface interface {
-	GetApiVersion() string
-	GetKind() string
-	GetMetadata() ApiObjectMetadataIface
-}
-
 // Options for defining API objects.
 // Experimental.
-// Struct proxy
 type ApiObjectProps struct {
 	// API version.
 	// Experimental.
@@ -494,144 +345,51 @@ type ApiObjectProps struct {
 	// If `name` is not specified, an app-unique name will be allocated by the
 	// framework based on the path of the construct within thes construct tree.
 	// Experimental.
-	Metadata ApiObjectMetadataIface `json:"metadata"`
-}
-
-func (a *ApiObjectProps) GetApiVersion() string {
-	var returns string
-	_jsii_.Get(
-		a,
-		"apiVersion",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (a *ApiObjectProps) GetKind() string {
-	var returns string
-	_jsii_.Get(
-		a,
-		"kind",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (a *ApiObjectProps) GetMetadata() ApiObjectMetadataIface {
-	var returns ApiObjectMetadataIface
-	_jsii_.Get(
-		a,
-		"metadata",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*ApiObjectMetadataIface)(nil)).Elem(): reflect.TypeOf((*ApiObjectMetadata)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-
-// Class interface
-type AppIface interface {
-	constructs.IConstructIface
-	GetOutdir() string
-	OnPrepare()
-	OnSynthesize(session constructs.ISynthesisSessionIface)
-	OnValidate() []string
-	ToString() string
-	Synth()
+	Metadata ApiObjectMetadata `json:"metadata"`
 }
 
 // Represents a cdk8s application.
 // Experimental.
-// Struct proxy
-type App struct {
-	// The output directory into which manifests will be synthesized.
-	// Experimental.
-	Outdir string `json:"outdir"`
+type App interface {
+	constructs.Construct
+	Outdir() string
+	Synth()
 }
 
-func (a *App) GetOutdir() string {
+// The jsii proxy struct for App
+type app struct {
+	constructs.Construct // extends constructs.Construct
+}
+
+func (a *app) Outdir() string {
 	var returns string
 	_jsii_.Get(
 		a,
 		"outdir",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
 
 // Defines an app.
-func NewApp(props AppPropsIface) AppIface {
+func NewApp(props AppProps) App {
 	_init_.Initialize()
-	self := App{}
+	a := app{}
+
 	_jsii_.Create(
 		"cdk8s.App",
 		[]interface{}{props},
 		[]_jsii_.FQN{},
 		[]_jsii_.Override{},
-		&self,
+		&a,
 	)
-	return &self
+	return &a
 }
 
-func (a *App) OnPrepare() {
-	var returns interface{}
-	_jsii_.Invoke(
-		a,
-		"onPrepare",
-		[]interface{}{},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (a *App) OnSynthesize(session constructs.ISynthesisSessionIface) {
-	var returns interface{}
-	_jsii_.Invoke(
-		a,
-		"onSynthesize",
-		[]interface{}{session},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (a *App) OnValidate() []string {
-	var returns []string
-	_jsii_.Invoke(
-		a,
-		"onValidate",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (a *App) ToString() string {
-	var returns string
-	_jsii_.Invoke(
-		a,
-		"toString",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (a *App) Synth() {
+// Synthesizes all manifests to the output directory.
+// Experimental.
+func (a *app) Synth() {
 	var returns interface{}
 	_jsii_.Invoke(
 		a,
@@ -639,170 +397,86 @@ func (a *App) Synth() {
 		[]interface{}{},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
-// AppPropsIface is the public interface for the custom type AppProps
 // Experimental.
-type AppPropsIface interface {
-	GetOutdir() string
-}
-
-// Experimental.
-// Struct proxy
 type AppProps struct {
 	// The directory to output Kubernetes manifests.
 	// Experimental.
 	Outdir string `json:"outdir"`
 }
 
-func (a *AppProps) GetOutdir() string {
-	var returns string
-	_jsii_.Get(
-		a,
-		"outdir",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-
-// Class interface
-type ChartIface interface {
-	constructs.IConstructIface
-	GetLabels() map[string]string
-	GetNamespace() string
-	OnPrepare()
-	OnSynthesize(session constructs.ISynthesisSessionIface)
-	OnValidate() []string
-	ToString() string
-	AddDependency(dependencies constructs.IConstructIface)
-	GenerateObjectName(apiObject ApiObjectIface) string
+// Experimental.
+type Chart interface {
+	constructs.Construct
+	Labels() map[string]string
+	Namespace() string
+	AddDependency(dependencies constructs.IConstruct)
+	GenerateObjectName(apiObject ApiObject) string
 	ToJson() []interface{}
 }
 
-// Experimental.
-// Struct proxy
-type Chart struct {
-	// Labels applied to all resources in this chart.
-	//
-	// This is an immutable copy.
-	// Experimental.
-	Labels map[string]string `json:"labels"`
-	// The default namespace for all objects in this chart.
-	// Experimental.
-	Namespace string `json:"namespace"`
+// The jsii proxy struct for Chart
+type chart struct {
+	constructs.Construct // extends constructs.Construct
 }
 
-func (c *Chart) GetLabels() map[string]string {
+func (c *chart) Labels() map[string]string {
 	var returns map[string]string
 	_jsii_.Get(
 		c,
 		"labels",
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (c *Chart) GetNamespace() string {
+func (c *chart) Namespace() string {
 	var returns string
 	_jsii_.Get(
 		c,
 		"namespace",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
 
-func NewChart(scope constructs.ConstructIface, id string, props ChartPropsIface) ChartIface {
+func NewChart(scope constructs.Construct, id string, props ChartProps) Chart {
 	_init_.Initialize()
-	self := Chart{}
+	c := chart{}
+
 	_jsii_.Create(
 		"cdk8s.Chart",
 		[]interface{}{scope, id, props},
 		[]_jsii_.FQN{},
 		[]_jsii_.Override{},
-		&self,
+		&c,
 	)
-	return &self
+	return &c
 }
 
-func Chart_Of(c constructs.IConstructIface) ChartIface {
+// Finds the chart in which a node is defined.
+// Experimental.
+func Chart_Of(c constructs.IConstruct) Chart {
 	_init_.Initialize()
-	var returns ChartIface
-	_jsii_.InvokeStatic(
+	var returns Chart
+	_jsii_.StaticInvoke(
 		"cdk8s.Chart",
 		"of",
 		[]interface{}{c},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*ChartIface)(nil)).Elem(): reflect.TypeOf((*Chart)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (c *Chart) OnPrepare() {
-	var returns interface{}
-	_jsii_.Invoke(
-		c,
-		"onPrepare",
-		[]interface{}{},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (c *Chart) OnSynthesize(session constructs.ISynthesisSessionIface) {
-	var returns interface{}
-	_jsii_.Invoke(
-		c,
-		"onSynthesize",
-		[]interface{}{session},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (c *Chart) OnValidate() []string {
-	var returns []string
-	_jsii_.Invoke(
-		c,
-		"onValidate",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (c *Chart) ToString() string {
-	var returns string
-	_jsii_.Invoke(
-		c,
-		"toString",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (c *Chart) AddDependency(dependencies constructs.IConstructIface) {
+// Create a dependency between this Chart and other constructs.
+//
+// These can be other ApiObjects, Charts, or custom.
+// Experimental.
+func (c *chart) AddDependency(dependencies constructs.IConstruct) {
 	var returns interface{}
 	_jsii_.Invoke(
 		c,
@@ -810,11 +484,27 @@ func (c *Chart) AddDependency(dependencies constructs.IConstructIface) {
 		[]interface{}{dependencies},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
-func (c *Chart) GenerateObjectName(apiObject ApiObjectIface) string {
+// Generates a app-unique name for an object given it's construct node path.
+//
+// Different resource types may have different constraints on names
+// (`metadata.name`). The previous version of the name generator was
+// compatible with DNS_SUBDOMAIN but not with DNS_LABEL.
+//
+// For example, `Deployment` names must comply with DNS_SUBDOMAIN while
+// `Service` names must comply with DNS_LABEL.
+//
+// Since there is no formal specification for this, the default name
+// generation scheme for kubernetes objects in cdk8s was changed to DNS_LABEL,
+// since it’s the common denominator for all kubernetes resources
+// (supposedly).
+//
+// You can override this method if you wish to customize object names at the
+// chart level.
+// Experimental.
+func (c *chart) GenerateObjectName(apiObject ApiObject) string {
 	var returns string
 	_jsii_.Invoke(
 		c,
@@ -822,12 +512,15 @@ func (c *Chart) GenerateObjectName(apiObject ApiObjectIface) string {
 		[]interface{}{apiObject},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (c *Chart) ToJson() []interface{} {
+// Renders this chart to a set of Kubernetes JSON resources.
+//
+// Returns: array of resource manifests
+// Experimental.
+func (c *chart) ToJson() []interface{} {
 	var returns []interface{}
 	_jsii_.Invoke(
 		c,
@@ -835,22 +528,11 @@ func (c *Chart) ToJson() []interface{} {
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*interface{})(nil)).Elem(): reflect.TypeOf((*interface{})(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-// ChartPropsIface is the public interface for the custom type ChartProps
 // Experimental.
-type ChartPropsIface interface {
-	GetLabels() map[string]string
-	GetNamespace() string
-}
-
-// Experimental.
-// Struct proxy
 type ChartProps struct {
 	// Labels to apply to all resources in this chart.
 	// Experimental.
@@ -861,37 +543,6 @@ type ChartProps struct {
 	// `namespace` explicitly defined for them.
 	// Experimental.
 	Namespace string `json:"namespace"`
-}
-
-func (c *ChartProps) GetLabels() map[string]string {
-	var returns map[string]string
-	_jsii_.Get(
-		c,
-		"labels",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (c *ChartProps) GetNamespace() string {
-	var returns string
-	_jsii_.Get(
-		c,
-		"namespace",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-
-// Class interface
-type DependencyGraphIface interface {
-	GetRoot() DependencyVertexIface
-	Topology() []constructs.IConstructIface
 }
 
 // Represents the dependency graph for a given Node.
@@ -905,140 +556,125 @@ type DependencyGraphIface interface {
 // The graph does not include cross-scope dependencies. That is, if a child on the current scope depends on a node
 // from a different scope, that relationship is not represented in this graph.
 // Experimental.
-// Struct proxy
-type DependencyGraph struct {
-	// Returns the root of the graph.
-	//
-	// Note that this vertex will always have `null` as its `.value` since it is an artifical root
-	// that binds all the connected spaces of the graph.
-	// Experimental.
-	Root DependencyVertexIface `json:"root"`
+type DependencyGraph interface {
+	Root() DependencyVertex
+	Topology() []constructs.IConstruct
 }
 
-func (d *DependencyGraph) GetRoot() DependencyVertexIface {
-	var returns DependencyVertexIface
+// The jsii proxy struct for DependencyGraph
+type dependencyGraph struct {
+	_ byte // padding
+}
+
+func (d *dependencyGraph) Root() DependencyVertex {
+	var returns DependencyVertex
 	_jsii_.Get(
 		d,
 		"root",
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DependencyVertexIface)(nil)).Elem(): reflect.TypeOf((*DependencyVertex)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
 
-func NewDependencyGraph(node constructs.NodeIface) DependencyGraphIface {
+func NewDependencyGraph(node constructs.Node) DependencyGraph {
 	_init_.Initialize()
-	self := DependencyGraph{}
+	d := dependencyGraph{}
+
 	_jsii_.Create(
 		"cdk8s.DependencyGraph",
 		[]interface{}{node},
 		[]_jsii_.FQN{},
 		[]_jsii_.Override{},
-		&self,
+		&d,
 	)
-	return &self
+	return &d
 }
 
-func (d *DependencyGraph) Topology() []constructs.IConstructIface {
-	var returns []constructs.IConstructIface
+// See: Vertex.topology()
+//
+// Experimental.
+func (d *dependencyGraph) Topology() []constructs.IConstruct {
+	var returns []constructs.IConstruct
 	_jsii_.Invoke(
 		d,
 		"topology",
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*constructs.IConstructIface)(nil)).Elem(): reflect.TypeOf((*constructs.IConstruct)(nil)).Elem(),
-		},
 	)
 	return returns
-}
-
-// Class interface
-type DependencyVertexIface interface {
-	GetInbound() []DependencyVertexIface
-	GetOutbound() []DependencyVertexIface
-	GetValue() constructs.IConstructIface
-	AddChild(dep DependencyVertexIface)
-	Topology() []constructs.IConstructIface
 }
 
 // Represents a vertex in the graph.
 //
 // The value of each vertex is an `IConstruct` that is accessible via the `.value` getter.
 // Experimental.
-// Struct proxy
-type DependencyVertex struct {
-	// Returns the parents of the vertex (i.e dependants).
-	// Experimental.
-	Inbound []DependencyVertexIface `json:"inbound"`
-	// Returns the children of the vertex (i.e dependencies).
-	// Experimental.
-	Outbound []DependencyVertexIface `json:"outbound"`
-	// Returns the IConstruct this graph vertex represents.
-	//
-	// `null` in case this is the root of the graph.
-	// Experimental.
-	Value constructs.IConstructIface `json:"value"`
+type DependencyVertex interface {
+	Inbound() []DependencyVertex
+	Outbound() []DependencyVertex
+	Value() constructs.IConstruct
+	AddChild(dep DependencyVertex)
+	Topology() []constructs.IConstruct
 }
 
-func (d *DependencyVertex) GetInbound() []DependencyVertexIface {
-	var returns []DependencyVertexIface
+// The jsii proxy struct for DependencyVertex
+type dependencyVertex struct {
+	_ byte // padding
+}
+
+func (d *dependencyVertex) Inbound() []DependencyVertex {
+	var returns []DependencyVertex
 	_jsii_.Get(
 		d,
 		"inbound",
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DependencyVertexIface)(nil)).Elem(): reflect.TypeOf((*DependencyVertex)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (d *DependencyVertex) GetOutbound() []DependencyVertexIface {
-	var returns []DependencyVertexIface
+func (d *dependencyVertex) Outbound() []DependencyVertex {
+	var returns []DependencyVertex
 	_jsii_.Get(
 		d,
 		"outbound",
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DependencyVertexIface)(nil)).Elem(): reflect.TypeOf((*DependencyVertex)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (d *DependencyVertex) GetValue() constructs.IConstructIface {
-	var returns constructs.IConstructIface
+func (d *dependencyVertex) Value() constructs.IConstruct {
+	var returns constructs.IConstruct
 	_jsii_.Get(
 		d,
 		"value",
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*constructs.IConstructIface)(nil)).Elem(): reflect.TypeOf((*constructs.IConstruct)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
 
-func NewDependencyVertex(value constructs.IConstructIface) DependencyVertexIface {
+func NewDependencyVertex(value constructs.IConstruct) DependencyVertex {
 	_init_.Initialize()
-	self := DependencyVertex{}
+	d := dependencyVertex{}
+
 	_jsii_.Create(
 		"cdk8s.DependencyVertex",
 		[]interface{}{value},
 		[]_jsii_.FQN{},
 		[]_jsii_.Override{},
-		&self,
+		&d,
 	)
-	return &self
+	return &d
 }
 
-func (d *DependencyVertex) AddChild(dep DependencyVertexIface) {
+// Adds a vertex as a dependency of the current node.
+//
+// Also updates the parents of `dep`, so that it contains this node as a parent.
+//
+// This operation will fail in case it creates a cycle in the graph.
+// Experimental.
+func (d *dependencyVertex) AddChild(dep DependencyVertex) {
 	var returns interface{}
 	_jsii_.Invoke(
 		d,
@@ -1046,34 +682,21 @@ func (d *DependencyVertex) AddChild(dep DependencyVertexIface) {
 		[]interface{}{dep},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
-func (d *DependencyVertex) Topology() []constructs.IConstructIface {
-	var returns []constructs.IConstructIface
+// Returns a topologically sorted array of the constructs in the sub-graph.
+// Experimental.
+func (d *dependencyVertex) Topology() []constructs.IConstruct {
+	var returns []constructs.IConstruct
 	_jsii_.Invoke(
 		d,
 		"topology",
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*constructs.IConstructIface)(nil)).Elem(): reflect.TypeOf((*constructs.IConstruct)(nil)).Elem(),
-		},
 	)
 	return returns
-}
-
-// Class interface
-type DurationIface interface {
-	ToDays(opts TimeConversionOptionsIface) float64
-	ToHours(opts TimeConversionOptionsIface) float64
-	ToHumanString() string
-	ToIsoString() string
-	ToMilliseconds(opts TimeConversionOptionsIface) float64
-	ToMinutes(opts TimeConversionOptionsIface) float64
-	ToSeconds(opts TimeConversionOptionsIface) float64
 }
 
 // Represents a length of time.
@@ -1081,107 +704,130 @@ type DurationIface interface {
 // The amount can be specified either as a literal value (e.g: `10`) which
 // cannot be negative.
 // Experimental.
-// Struct proxy
-type Duration struct {
+type Duration interface {
+	ToDays(opts TimeConversionOptions) float64
+	ToHours(opts TimeConversionOptions) float64
+	ToHumanString() string
+	ToIsoString() string
+	ToMilliseconds(opts TimeConversionOptions) float64
+	ToMinutes(opts TimeConversionOptions) float64
+	ToSeconds(opts TimeConversionOptions) float64
 }
 
-func Duration_Days(amount float64) DurationIface {
+// The jsii proxy struct for Duration
+type duration struct {
+	_ byte // padding
+}
+
+// Create a Duration representing an amount of days.
+//
+// Returns: a new `Duration` representing `amount` Days.
+// Experimental.
+func Duration_Days(amount float64) Duration {
 	_init_.Initialize()
-	var returns DurationIface
-	_jsii_.InvokeStatic(
+	var returns Duration
+	_jsii_.StaticInvoke(
 		"cdk8s.Duration",
 		"days",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DurationIface)(nil)).Elem(): reflect.TypeOf((*Duration)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Duration_Hours(amount float64) DurationIface {
+// Create a Duration representing an amount of hours.
+//
+// Returns: a new `Duration` representing `amount` Hours.
+// Experimental.
+func Duration_Hours(amount float64) Duration {
 	_init_.Initialize()
-	var returns DurationIface
-	_jsii_.InvokeStatic(
+	var returns Duration
+	_jsii_.StaticInvoke(
 		"cdk8s.Duration",
 		"hours",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DurationIface)(nil)).Elem(): reflect.TypeOf((*Duration)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Duration_Millis(amount float64) DurationIface {
+// Create a Duration representing an amount of milliseconds.
+//
+// Returns: a new `Duration` representing `amount` ms.
+// Experimental.
+func Duration_Millis(amount float64) Duration {
 	_init_.Initialize()
-	var returns DurationIface
-	_jsii_.InvokeStatic(
+	var returns Duration
+	_jsii_.StaticInvoke(
 		"cdk8s.Duration",
 		"millis",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DurationIface)(nil)).Elem(): reflect.TypeOf((*Duration)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Duration_Minutes(amount float64) DurationIface {
+// Create a Duration representing an amount of minutes.
+//
+// Returns: a new `Duration` representing `amount` Minutes.
+// Experimental.
+func Duration_Minutes(amount float64) Duration {
 	_init_.Initialize()
-	var returns DurationIface
-	_jsii_.InvokeStatic(
+	var returns Duration
+	_jsii_.StaticInvoke(
 		"cdk8s.Duration",
 		"minutes",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DurationIface)(nil)).Elem(): reflect.TypeOf((*Duration)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Duration_Parse(duration string) DurationIface {
+// Parse a period formatted according to the ISO 8601 standard.
+//
+// Returns: the parsed `Duration`.
+// See: https://www.iso.org/fr/standard/70907.html
+//
+// Experimental.
+func Duration_Parse(duration string) Duration {
 	_init_.Initialize()
-	var returns DurationIface
-	_jsii_.InvokeStatic(
+	var returns Duration
+	_jsii_.StaticInvoke(
 		"cdk8s.Duration",
 		"parse",
 		[]interface{}{duration},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DurationIface)(nil)).Elem(): reflect.TypeOf((*Duration)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Duration_Seconds(amount float64) DurationIface {
+// Create a Duration representing an amount of seconds.
+//
+// Returns: a new `Duration` representing `amount` Seconds.
+// Experimental.
+func Duration_Seconds(amount float64) Duration {
 	_init_.Initialize()
-	var returns DurationIface
-	_jsii_.InvokeStatic(
+	var returns Duration
+	_jsii_.StaticInvoke(
 		"cdk8s.Duration",
 		"seconds",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*DurationIface)(nil)).Elem(): reflect.TypeOf((*Duration)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (d *Duration) ToDays(opts TimeConversionOptionsIface) float64 {
+// Return the total number of days in this Duration.
+//
+// Returns: the value of this `Duration` expressed in Days.
+// Experimental.
+func (d *duration) ToDays(opts TimeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		d,
@@ -1189,12 +835,15 @@ func (d *Duration) ToDays(opts TimeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (d *Duration) ToHours(opts TimeConversionOptionsIface) float64 {
+// Return the total number of hours in this Duration.
+//
+// Returns: the value of this `Duration` expressed in Hours.
+// Experimental.
+func (d *duration) ToHours(opts TimeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		d,
@@ -1202,12 +851,13 @@ func (d *Duration) ToHours(opts TimeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (d *Duration) ToHumanString() string {
+// Turn this duration into a human-readable string.
+// Experimental.
+func (d *duration) ToHumanString() string {
 	var returns string
 	_jsii_.Invoke(
 		d,
@@ -1215,12 +865,17 @@ func (d *Duration) ToHumanString() string {
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (d *Duration) ToIsoString() string {
+// Return an ISO 8601 representation of this period.
+//
+// Returns: a string starting with 'PT' describing the period
+// See: https://www.iso.org/fr/standard/70907.html
+//
+// Experimental.
+func (d *duration) ToIsoString() string {
 	var returns string
 	_jsii_.Invoke(
 		d,
@@ -1228,12 +883,15 @@ func (d *Duration) ToIsoString() string {
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (d *Duration) ToMilliseconds(opts TimeConversionOptionsIface) float64 {
+// Return the total number of milliseconds in this Duration.
+//
+// Returns: the value of this `Duration` expressed in Milliseconds.
+// Experimental.
+func (d *duration) ToMilliseconds(opts TimeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		d,
@@ -1241,12 +899,15 @@ func (d *Duration) ToMilliseconds(opts TimeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (d *Duration) ToMinutes(opts TimeConversionOptionsIface) float64 {
+// Return the total number of minutes in this Duration.
+//
+// Returns: the value of this `Duration` expressed in Minutes.
+// Experimental.
+func (d *duration) ToMinutes(opts TimeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		d,
@@ -1254,12 +915,15 @@ func (d *Duration) ToMinutes(opts TimeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (d *Duration) ToSeconds(opts TimeConversionOptionsIface) float64 {
+// Return the total number of seconds in this Duration.
+//
+// Returns: the value of this `Duration` expressed in Seconds.
+// Experimental.
+func (d *duration) ToSeconds(opts TimeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		d,
@@ -1267,20 +931,11 @@ func (d *Duration) ToSeconds(opts TimeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-// GroupVersionKindIface is the public interface for the custom type GroupVersionKind
 // Experimental.
-type GroupVersionKindIface interface {
-	GetApiVersion() string
-	GetKind() string
-}
-
-// Experimental.
-// Struct proxy
 type GroupVersionKind struct {
 	// The object's API version (e.g. `authorization.k8s.io/v1`).
 	// Experimental.
@@ -1290,157 +945,47 @@ type GroupVersionKind struct {
 	Kind string `json:"kind"`
 }
 
-func (g *GroupVersionKind) GetApiVersion() string {
-	var returns string
-	_jsii_.Get(
-		g,
-		"apiVersion",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (g *GroupVersionKind) GetKind() string {
-	var returns string
-	_jsii_.Get(
-		g,
-		"kind",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-
-// Class interface
-type HelmIface interface {
-	constructs.IConstructIface
-	GetApiObjects() []ApiObjectIface
-	GetReleaseName() string
-	OnPrepare()
-	OnSynthesize(session constructs.ISynthesisSessionIface)
-	OnValidate() []string
-	ToString() string
-}
-
 // Represents a Helm deployment.
 //
 // Use this construct to import an existing Helm chart and incorporate it into your constructs.
 // Experimental.
-// Struct proxy
-type Helm struct {
-	// Returns all the included API objects.
-	// Experimental.
-	ApiObjects []ApiObjectIface `json:"apiObjects"`
-	// The helm release name.
-	// Experimental.
-	ReleaseName string `json:"releaseName"`
+type Helm interface {
+	Include
+	ReleaseName() string
 }
 
-func (h *Helm) GetApiObjects() []ApiObjectIface {
-	var returns []ApiObjectIface
-	_jsii_.Get(
-		h,
-		"apiObjects",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*ApiObjectIface)(nil)).Elem(): reflect.TypeOf((*ApiObject)(nil)).Elem(),
-		},
-	)
-	return returns
+// The jsii proxy struct for Helm
+type helm struct {
+	include // extends cdk8s.Include
 }
 
-func (h *Helm) GetReleaseName() string {
+func (h *helm) ReleaseName() string {
 	var returns string
 	_jsii_.Get(
 		h,
 		"releaseName",
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
 
-func NewHelm(scope constructs.ConstructIface, id string, props HelmPropsIface) HelmIface {
+func NewHelm(scope constructs.Construct, id string, props HelmProps) Helm {
 	_init_.Initialize()
-	self := Helm{}
+	h := helm{}
+
 	_jsii_.Create(
 		"cdk8s.Helm",
 		[]interface{}{scope, id, props},
 		[]_jsii_.FQN{},
 		[]_jsii_.Override{},
-		&self,
+		&h,
 	)
-	return &self
-}
-
-func (h *Helm) OnPrepare() {
-	var returns interface{}
-	_jsii_.Invoke(
-		h,
-		"onPrepare",
-		[]interface{}{},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (h *Helm) OnSynthesize(session constructs.ISynthesisSessionIface) {
-	var returns interface{}
-	_jsii_.Invoke(
-		h,
-		"onSynthesize",
-		[]interface{}{session},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (h *Helm) OnValidate() []string {
-	var returns []string
-	_jsii_.Invoke(
-		h,
-		"onValidate",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (h *Helm) ToString() string {
-	var returns string
-	_jsii_.Invoke(
-		h,
-		"toString",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-// HelmPropsIface is the public interface for the custom type HelmProps
-// Experimental.
-type HelmPropsIface interface {
-	GetChart() string
-	GetHelmExecutable() string
-	GetHelmFlags() []string
-	GetReleaseName() string
-	GetValues() map[string]interface{}
+	return &h
 }
 
 // Options for `Helm`.
 // Experimental.
-// Struct proxy
 type HelmProps struct {
 	// The chart name to use. It can be a chart from a helm repository or a local directory.
 	//
@@ -1466,75 +1011,18 @@ type HelmProps struct {
 	Values map[string]interface{} `json:"values"`
 }
 
-func (h *HelmProps) GetChart() string {
-	var returns string
-	_jsii_.Get(
-		h,
-		"chart",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (h *HelmProps) GetHelmExecutable() string {
-	var returns string
-	_jsii_.Get(
-		h,
-		"helmExecutable",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (h *HelmProps) GetHelmFlags() []string {
-	var returns []string
-	_jsii_.Get(
-		h,
-		"helmFlags",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (h *HelmProps) GetReleaseName() string {
-	var returns string
-	_jsii_.Get(
-		h,
-		"releaseName",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (h *HelmProps) GetValues() map[string]interface{} {
-	var returns map[string]interface{}
-	_jsii_.Get(
-		h,
-		"values",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*interface{})(nil)).Elem(): reflect.TypeOf((*interface{})(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-
 // Experimental.
-type IAnyProducerIface interface {
+type IAnyProducer interface {
 	// Experimental.
 	Produce() interface{}
 }
 
-type IAnyProducer struct {}
+// The jsii proxy for IAnyProducer
+type iAnyProducer struct {
+	_ byte // padding
+}
 
-func (i *IAnyProducer) Produce() interface{} {
+func (i *iAnyProducer) Produce() interface{} {
 	var returns interface{}
 	_jsii_.Invoke(
 		i,
@@ -1542,19 +1030,8 @@ func (i *IAnyProducer) Produce() interface{} {
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
-}
-
-// Class interface
-type IncludeIface interface {
-	constructs.IConstructIface
-	GetApiObjects() []ApiObjectIface
-	OnPrepare()
-	OnSynthesize(session constructs.ISynthesisSessionIface)
-	OnValidate() []string
-	ToString() string
 }
 
 // Reads a YAML manifest from a file or a URL and defines all resources as API objects within the defined scope.
@@ -1562,100 +1039,42 @@ type IncludeIface interface {
 // The names (`metadata.name`) of imported resources will be preserved as-is
 // from the manifest.
 // Experimental.
-// Struct proxy
-type Include struct {
-	// Returns all the included API objects.
-	// Experimental.
-	ApiObjects []ApiObjectIface `json:"apiObjects"`
+type Include interface {
+	constructs.Construct
+	ApiObjects() []ApiObject
 }
 
-func (i *Include) GetApiObjects() []ApiObjectIface {
-	var returns []ApiObjectIface
+// The jsii proxy struct for Include
+type include struct {
+	constructs.Construct // extends constructs.Construct
+}
+
+func (i *include) ApiObjects() []ApiObject {
+	var returns []ApiObject
 	_jsii_.Get(
 		i,
 		"apiObjects",
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*ApiObjectIface)(nil)).Elem(): reflect.TypeOf((*ApiObject)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
 
-func NewInclude(scope constructs.ConstructIface, id string, props IncludePropsIface) IncludeIface {
+func NewInclude(scope constructs.Construct, id string, props IncludeProps) Include {
 	_init_.Initialize()
-	self := Include{}
+	i := include{}
+
 	_jsii_.Create(
 		"cdk8s.Include",
 		[]interface{}{scope, id, props},
 		[]_jsii_.FQN{},
 		[]_jsii_.Override{},
-		&self,
+		&i,
 	)
-	return &self
-}
-
-func (i *Include) OnPrepare() {
-	var returns interface{}
-	_jsii_.Invoke(
-		i,
-		"onPrepare",
-		[]interface{}{},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (i *Include) OnSynthesize(session constructs.ISynthesisSessionIface) {
-	var returns interface{}
-	_jsii_.Invoke(
-		i,
-		"onSynthesize",
-		[]interface{}{session},
-		false,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-}
-
-func (i *Include) OnValidate() []string {
-	var returns []string
-	_jsii_.Invoke(
-		i,
-		"onValidate",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (i *Include) ToString() string {
-	var returns string
-	_jsii_.Invoke(
-		i,
-		"toString",
-		[]interface{}{},
-		true,
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-// IncludePropsIface is the public interface for the custom type IncludeProps
-// Experimental.
-type IncludePropsIface interface {
-	GetUrl() string
+	return &i
 }
 
 // Experimental.
-// Struct proxy
 type IncludeProps struct {
 	// Local file path or URL which includes a Kubernetes YAML manifest.
 	//
@@ -1663,22 +1082,6 @@ type IncludeProps struct {
 	//
 	// Experimental.
 	Url string `json:"url"`
-}
-
-func (i *IncludeProps) GetUrl() string {
-	var returns string
-	_jsii_.Get(
-		i,
-		"url",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-
-// Class interface
-type JsonPatchIface interface {
 }
 
 // Utility for applying RFC-6902 JSON-Patch to a document.
@@ -1692,145 +1095,179 @@ type JsonPatchIface interface {
 // TODO: EXAMPLE
 //
 // Experimental.
-// Struct proxy
-type JsonPatch struct {
+type JsonPatch interface {
 }
 
-func JsonPatch_Add(path string, value interface{}) JsonPatchIface {
+// The jsii proxy struct for JsonPatch
+type jsonPatch struct {
+	_ byte // padding
+}
+
+// Adds a value to an object or inserts it into an array.
+//
+// In the case of an
+// array, the value is inserted before the given index. The - character can be
+// used instead of an index to insert at the end of an array.
+//
+// TODO: EXAMPLE
+//
+// Experimental.
+func JsonPatch_Add(path string, value interface{}) JsonPatch {
 	_init_.Initialize()
-	var returns JsonPatchIface
-	_jsii_.InvokeStatic(
+	var returns JsonPatch
+	_jsii_.StaticInvoke(
 		"cdk8s.JsonPatch",
 		"add",
 		[]interface{}{path, value},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*JsonPatchIface)(nil)).Elem(): reflect.TypeOf((*JsonPatch)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func JsonPatch_Apply(document interface{}, ops JsonPatchIface) interface{} {
+// Applies a set of JSON-Patch (RFC-6902) operations to `document` and returns the result.
+//
+// Returns: The result document
+// Experimental.
+func JsonPatch_Apply(document interface{}, ops JsonPatch) interface{} {
 	_init_.Initialize()
 	var returns interface{}
-	_jsii_.InvokeStatic(
+	_jsii_.StaticInvoke(
 		"cdk8s.JsonPatch",
 		"apply",
 		[]interface{}{document, ops},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func JsonPatch_Copy(from string, path string) JsonPatchIface {
+// Copies a value from one location to another within the JSON document.
+//
+// Both
+// from and path are JSON Pointers.
+//
+// TODO: EXAMPLE
+//
+// Experimental.
+func JsonPatch_Copy(from string, path string) JsonPatch {
 	_init_.Initialize()
-	var returns JsonPatchIface
-	_jsii_.InvokeStatic(
+	var returns JsonPatch
+	_jsii_.StaticInvoke(
 		"cdk8s.JsonPatch",
 		"copy",
 		[]interface{}{from, path},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*JsonPatchIface)(nil)).Elem(): reflect.TypeOf((*JsonPatch)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func JsonPatch_Move(from string, path string) JsonPatchIface {
+// Moves a value from one location to the other.
+//
+// Both from and path are JSON Pointers.
+//
+// TODO: EXAMPLE
+//
+// Experimental.
+func JsonPatch_Move(from string, path string) JsonPatch {
 	_init_.Initialize()
-	var returns JsonPatchIface
-	_jsii_.InvokeStatic(
+	var returns JsonPatch
+	_jsii_.StaticInvoke(
 		"cdk8s.JsonPatch",
 		"move",
 		[]interface{}{from, path},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*JsonPatchIface)(nil)).Elem(): reflect.TypeOf((*JsonPatch)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func JsonPatch_Remove(path string) JsonPatchIface {
+// Removes a value from an object or array.
+//
+// TODO: EXAMPLE
+//
+// Experimental.
+func JsonPatch_Remove(path string) JsonPatch {
 	_init_.Initialize()
-	var returns JsonPatchIface
-	_jsii_.InvokeStatic(
+	var returns JsonPatch
+	_jsii_.StaticInvoke(
 		"cdk8s.JsonPatch",
 		"remove",
 		[]interface{}{path},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*JsonPatchIface)(nil)).Elem(): reflect.TypeOf((*JsonPatch)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func JsonPatch_Replace(path string, value interface{}) JsonPatchIface {
+// Replaces a value.
+//
+// Equivalent to a “remove” followed by an “add”.
+//
+// TODO: EXAMPLE
+//
+// Experimental.
+func JsonPatch_Replace(path string, value interface{}) JsonPatch {
 	_init_.Initialize()
-	var returns JsonPatchIface
-	_jsii_.InvokeStatic(
+	var returns JsonPatch
+	_jsii_.StaticInvoke(
 		"cdk8s.JsonPatch",
 		"replace",
 		[]interface{}{path, value},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*JsonPatchIface)(nil)).Elem(): reflect.TypeOf((*JsonPatch)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func JsonPatch_Test(path string, value interface{}) JsonPatchIface {
+// Tests that the specified value is set in the document.
+//
+// If the test fails,
+// then the patch as a whole should not apply.
+//
+// TODO: EXAMPLE
+//
+// Experimental.
+func JsonPatch_Test(path string, value interface{}) JsonPatch {
 	_init_.Initialize()
-	var returns JsonPatchIface
-	_jsii_.InvokeStatic(
+	var returns JsonPatch
+	_jsii_.StaticInvoke(
 		"cdk8s.JsonPatch",
 		"test",
 		[]interface{}{path, value},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*JsonPatchIface)(nil)).Elem(): reflect.TypeOf((*JsonPatch)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-// Class interface
-type LazyIface interface {
+// Experimental.
+type Lazy interface {
 	Produce() interface{}
 }
 
-// Experimental.
-// Struct proxy
-type Lazy struct {
+// The jsii proxy struct for Lazy
+type lazy struct {
+	_ byte // padding
 }
 
-func Lazy_Any(producer IAnyProducerIface) interface{} {
+// Experimental.
+func Lazy_Any(producer IAnyProducer) interface{} {
 	_init_.Initialize()
 	var returns interface{}
-	_jsii_.InvokeStatic(
+	_jsii_.StaticInvoke(
 		"cdk8s.Lazy",
 		"any",
 		[]interface{}{producer},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (l *Lazy) Produce() interface{} {
+// Experimental.
+func (l *lazy) Produce() interface{} {
 	var returns interface{}
 	_jsii_.Invoke(
 		l,
@@ -1838,23 +1275,12 @@ func (l *Lazy) Produce() interface{} {
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-// NameOptionsIface is the public interface for the custom type NameOptions
-// Experimental.
-type NameOptionsIface interface {
-	GetDelimiter() string
-	GetExtra() []string
-	GetIncludeHash() bool
-	GetMaxLen() float64
-}
-
 // Options for name generation.
 // Experimental.
-// Struct proxy
 type NameOptions struct {
 	// Delimiter to use between components.
 	// Experimental.
@@ -1870,98 +1296,76 @@ type NameOptions struct {
 	MaxLen float64 `json:"maxLen"`
 }
 
-func (n *NameOptions) GetDelimiter() string {
-	var returns string
-	_jsii_.Get(
-		n,
-		"delimiter",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (n *NameOptions) GetExtra() []string {
-	var returns []string
-	_jsii_.Get(
-		n,
-		"extra",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*string)(nil)).Elem(): reflect.TypeOf((*string)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-func (n *NameOptions) GetIncludeHash() bool {
-	var returns bool
-	_jsii_.Get(
-		n,
-		"includeHash",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-func (n *NameOptions) GetMaxLen() float64 {
-	var returns float64
-	_jsii_.Get(
-		n,
-		"maxLen",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-
-// Class interface
-type NamesIface interface {
-}
-
 // Utilities for generating unique and stable names.
 // Experimental.
-// Struct proxy
-type Names struct {
+type Names interface {
 }
 
-func Names_ToDnsLabel(scope constructs.ConstructIface, options NameOptionsIface) string {
+// The jsii proxy struct for Names
+type names struct {
+	_ byte // padding
+}
+
+// Generates a unique and stable name compatible DNS_LABEL from RFC-1123 from a path.
+//
+// The generated name will:
+//   - contain at most 63 characters
+//   - contain only lowercase alphanumeric characters or ‘-’
+//   - start with an alphanumeric character
+//   - end with an alphanumeric character
+//
+// The generated name will have the form:
+//   <comp0>-<comp1>-..-<compN>-<short-hash>
+//
+// Where <comp> are the path components (assuming they are is separated by
+// "/").
+//
+// Note that if the total length is longer than 63 characters, we will trim
+// the first components since the last components usually encode more meaning.
+// Experimental.
+func Names_ToDnsLabel(scope constructs.Construct, options NameOptions) string {
 	_init_.Initialize()
 	var returns string
-	_jsii_.InvokeStatic(
+	_jsii_.StaticInvoke(
 		"cdk8s.Names",
 		"toDnsLabel",
 		[]interface{}{scope, options},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func Names_ToLabelValue(scope constructs.ConstructIface, options NameOptionsIface) string {
+// Generates a unique and stable name compatible label key name segment and label value from a path.
+//
+// The name segment is required and must be 63 characters or less, beginning
+// and ending with an alphanumeric character ([a-z0-9A-Z]) with dashes (-),
+// underscores (_), dots (.), and alphanumerics between.
+//
+// Valid label values must be 63 characters or less and must be empty or
+// begin and end with an alphanumeric character ([a-z0-9A-Z]) with dashes
+// (-), underscores (_), dots (.), and alphanumerics between.
+//
+// The generated name will have the form:
+//   <comp0><delim><comp1><delim>..<delim><compN><delim><short-hash>
+//
+// Where <comp> are the path components (assuming they are is separated by
+// "/").
+//
+// Note that if the total length is longer than 63 characters, we will trim
+// the first components since the last components usually encode more meaning.
+// Experimental.
+func Names_ToLabelValue(scope constructs.Construct, options NameOptions) string {
 	_init_.Initialize()
 	var returns string
-	_jsii_.InvokeStatic(
+	_jsii_.StaticInvoke(
 		"cdk8s.Names",
 		"toLabelValue",
 		[]interface{}{scope, options},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
-}
-
-// Class interface
-type SizeIface interface {
-	ToGibibytes(opts SizeConversionOptionsIface) float64
-	ToKibibytes(opts SizeConversionOptionsIface) float64
-	ToMebibytes(opts SizeConversionOptionsIface) float64
-	ToPebibytes(opts SizeConversionOptionsIface) float64
-	ToTebibytes(opts SizeConversionOptionsIface) float64
 }
 
 // Represents the amount of digital storage.
@@ -1971,91 +1375,107 @@ type SizeIface interface {
 //
 // When the amount is passed as a token, unit conversion is not possible.
 // Experimental.
-// Struct proxy
-type Size struct {
+type Size interface {
+	ToGibibytes(opts SizeConversionOptions) float64
+	ToKibibytes(opts SizeConversionOptions) float64
+	ToMebibytes(opts SizeConversionOptions) float64
+	ToPebibytes(opts SizeConversionOptions) float64
+	ToTebibytes(opts SizeConversionOptions) float64
 }
 
-func Size_Gibibytes(amount float64) SizeIface {
+// The jsii proxy struct for Size
+type size struct {
+	_ byte // padding
+}
+
+// Create a Storage representing an amount gibibytes.
+//
+// 1 GiB = 1024 MiB
+// Experimental.
+func Size_Gibibytes(amount float64) Size {
 	_init_.Initialize()
-	var returns SizeIface
-	_jsii_.InvokeStatic(
+	var returns Size
+	_jsii_.StaticInvoke(
 		"cdk8s.Size",
 		"gibibytes",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*SizeIface)(nil)).Elem(): reflect.TypeOf((*Size)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Size_Kibibytes(amount float64) SizeIface {
+// Create a Storage representing an amount kibibytes.
+//
+// 1 KiB = 1024 bytes
+// Experimental.
+func Size_Kibibytes(amount float64) Size {
 	_init_.Initialize()
-	var returns SizeIface
-	_jsii_.InvokeStatic(
+	var returns Size
+	_jsii_.StaticInvoke(
 		"cdk8s.Size",
 		"kibibytes",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*SizeIface)(nil)).Elem(): reflect.TypeOf((*Size)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Size_Mebibytes(amount float64) SizeIface {
+// Create a Storage representing an amount mebibytes.
+//
+// 1 MiB = 1024 KiB
+// Experimental.
+func Size_Mebibytes(amount float64) Size {
 	_init_.Initialize()
-	var returns SizeIface
-	_jsii_.InvokeStatic(
+	var returns Size
+	_jsii_.StaticInvoke(
 		"cdk8s.Size",
 		"mebibytes",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*SizeIface)(nil)).Elem(): reflect.TypeOf((*Size)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Size_Pebibyte(amount float64) SizeIface {
+// Create a Storage representing an amount pebibytes.
+//
+// 1 PiB = 1024 TiB
+// Experimental.
+func Size_Pebibyte(amount float64) Size {
 	_init_.Initialize()
-	var returns SizeIface
-	_jsii_.InvokeStatic(
+	var returns Size
+	_jsii_.StaticInvoke(
 		"cdk8s.Size",
 		"pebibyte",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*SizeIface)(nil)).Elem(): reflect.TypeOf((*Size)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Size_Tebibytes(amount float64) SizeIface {
+// Create a Storage representing an amount tebibytes.
+//
+// 1 TiB = 1024 GiB
+// Experimental.
+func Size_Tebibytes(amount float64) Size {
 	_init_.Initialize()
-	var returns SizeIface
-	_jsii_.InvokeStatic(
+	var returns Size
+	_jsii_.StaticInvoke(
 		"cdk8s.Size",
 		"tebibytes",
 		[]interface{}{amount},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*SizeIface)(nil)).Elem(): reflect.TypeOf((*Size)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func (s *Size) ToGibibytes(opts SizeConversionOptionsIface) float64 {
+// Return this storage as a total number of gibibytes.
+// Experimental.
+func (s *size) ToGibibytes(opts SizeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		s,
@@ -2063,12 +1483,13 @@ func (s *Size) ToGibibytes(opts SizeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (s *Size) ToKibibytes(opts SizeConversionOptionsIface) float64 {
+// Return this storage as a total number of kibibytes.
+// Experimental.
+func (s *size) ToKibibytes(opts SizeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		s,
@@ -2076,12 +1497,13 @@ func (s *Size) ToKibibytes(opts SizeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (s *Size) ToMebibytes(opts SizeConversionOptionsIface) float64 {
+// Return this storage as a total number of mebibytes.
+// Experimental.
+func (s *size) ToMebibytes(opts SizeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		s,
@@ -2089,12 +1511,13 @@ func (s *Size) ToMebibytes(opts SizeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (s *Size) ToPebibytes(opts SizeConversionOptionsIface) float64 {
+// Return this storage as a total number of pebibytes.
+// Experimental.
+func (s *size) ToPebibytes(opts SizeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		s,
@@ -2102,12 +1525,13 @@ func (s *Size) ToPebibytes(opts SizeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-func (s *Size) ToTebibytes(opts SizeConversionOptionsIface) float64 {
+// Return this storage as a total number of tebibytes.
+// Experimental.
+func (s *size) ToTebibytes(opts SizeConversionOptions) float64 {
 	var returns float64
 	_jsii_.Invoke(
 		s,
@@ -2115,184 +1539,147 @@ func (s *Size) ToTebibytes(opts SizeConversionOptionsIface) float64 {
 		[]interface{}{opts},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
 
-// SizeConversionOptionsIface is the public interface for the custom type SizeConversionOptions
-// Experimental.
-type SizeConversionOptionsIface interface {
-	GetRounding() SizeRoundingBehavior
-}
-
 // Options for how to convert time to a different unit.
 // Experimental.
-// Struct proxy
 type SizeConversionOptions struct {
 	// How conversions should behave when it encounters a non-integer result.
 	// Experimental.
 	Rounding SizeRoundingBehavior `json:"rounding"`
 }
 
-func (s *SizeConversionOptions) GetRounding() SizeRoundingBehavior {
-	var returns SizeRoundingBehavior
-	_jsii_.Get(
-		s,
-		"rounding",
-		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*SizeRoundingBehavior)(nil)).Elem(): reflect.TypeOf((*SizeRoundingBehavior)(nil)).Elem(),
-		},
-	)
-	return returns
-}
-
-
 // Rounding behaviour when converting between units of `Size`.
 // Experimental.
 type SizeRoundingBehavior string
 
 const (
-	SizeRoundingBehaviorFail SizeRoundingBehavior = "FAIL"
-	SizeRoundingBehaviorFloor SizeRoundingBehavior = "FLOOR"
-	SizeRoundingBehaviorNone SizeRoundingBehavior = "NONE"
+	SizeRoundingBehavior_FAIL SizeRoundingBehavior = "FAIL"
+	SizeRoundingBehavior_FLOOR SizeRoundingBehavior = "FLOOR"
+	SizeRoundingBehavior_NONE SizeRoundingBehavior = "NONE"
 )
-
-// Class interface
-type TestingIface interface {
-}
 
 // Testing utilities for cdk8s applications.
 // Experimental.
-// Struct proxy
-type Testing struct {
+type Testing interface {
 }
 
-func Testing_App() AppIface {
+// The jsii proxy struct for Testing
+type testing struct {
+	_ byte // padding
+}
+
+// Returns an app for testing with the following properties: - Output directory is a temp dir.
+// Experimental.
+func Testing_App() App {
 	_init_.Initialize()
-	var returns AppIface
-	_jsii_.InvokeStatic(
+	var returns App
+	_jsii_.StaticInvoke(
 		"cdk8s.Testing",
 		"app",
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*AppIface)(nil)).Elem(): reflect.TypeOf((*App)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Testing_Chart() ChartIface {
+// Returns: a Chart that can be used for tests
+// Experimental.
+func Testing_Chart() Chart {
 	_init_.Initialize()
-	var returns ChartIface
-	_jsii_.InvokeStatic(
+	var returns Chart
+	_jsii_.StaticInvoke(
 		"cdk8s.Testing",
 		"chart",
 		[]interface{}{},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*ChartIface)(nil)).Elem(): reflect.TypeOf((*Chart)(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-func Testing_Synth(chart ChartIface) []interface{} {
+// Returns the Kubernetes manifest synthesized from this chart.
+// Experimental.
+func Testing_Synth(chart Chart) []interface{} {
 	_init_.Initialize()
 	var returns []interface{}
-	_jsii_.InvokeStatic(
+	_jsii_.StaticInvoke(
 		"cdk8s.Testing",
 		"synth",
 		[]interface{}{chart},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*interface{})(nil)).Elem(): reflect.TypeOf((*interface{})(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
-// TimeConversionOptionsIface is the public interface for the custom type TimeConversionOptions
-// Experimental.
-type TimeConversionOptionsIface interface {
-	GetIntegral() bool
-}
-
 // Options for how to convert time to a different unit.
 // Experimental.
-// Struct proxy
 type TimeConversionOptions struct {
 	// If `true`, conversions into a larger time unit (e.g. `Seconds` to `Minutes`) will fail if the result is not an integer.
 	// Experimental.
 	Integral bool `json:"integral"`
 }
 
-func (t *TimeConversionOptions) GetIntegral() bool {
-	var returns bool
-	_jsii_.Get(
-		t,
-		"integral",
-		&returns,
-		map[reflect.Type]reflect.Type{},
-	)
-	return returns
-}
-
-
-// Class interface
-type YamlIface interface {
-}
-
 // YAML utilities.
 // Experimental.
-// Struct proxy
-type Yaml struct {
+type Yaml interface {
 }
 
+// The jsii proxy struct for Yaml
+type yaml struct {
+	_ byte // padding
+}
+
+// Downloads a set of YAML documents (k8s manifest for example) from a URL or a file and returns them as javascript objects.
+//
+// Empty documents are filtered out.
+//
+// Returns: an array of objects, each represents a document inside the YAML
+// Experimental.
 func Yaml_Load(urlOrFile string) []interface{} {
 	_init_.Initialize()
 	var returns []interface{}
-	_jsii_.InvokeStatic(
+	_jsii_.StaticInvoke(
 		"cdk8s.Yaml",
 		"load",
 		[]interface{}{urlOrFile},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{
-			reflect.TypeOf((*interface{})(nil)).Elem(): reflect.TypeOf((*interface{})(nil)).Elem(),
-		},
 	)
 	return returns
 }
 
+// Saves a set of objects as a multi-document YAML file.
+// Experimental.
 func Yaml_Save(filePath string, docs []interface{}) {
 	_init_.Initialize()
 	var returns interface{}
-	_jsii_.InvokeStatic(
+	_jsii_.StaticInvoke(
 		"cdk8s.Yaml",
 		"save",
 		[]interface{}{filePath, docs},
 		false,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 }
 
+// Saves a set of YAML documents into a temp file (in /tmp).
+//
+// Returns: the path to the temporary file
+// Experimental.
 func Yaml_Tmp(docs []interface{}) string {
 	_init_.Initialize()
 	var returns string
-	_jsii_.InvokeStatic(
+	_jsii_.StaticInvoke(
 		"cdk8s.Yaml",
 		"tmp",
 		[]interface{}{docs},
 		true,
 		&returns,
-		map[reflect.Type]reflect.Type{},
 	)
 	return returns
 }
