@@ -1,13 +1,17 @@
 package jsii
 
 import (
-	rt "github.com/aws/jsii-runtime-go"
+	_          "embed"
 	"sync"
-	// Initialization endpoints of dependencies
+
+	_jsii_     "github.com/aws/jsii-runtime-go"
+
 	constructs "github.com/aws/constructs-go/constructs/v3/jsii"
 )
 
-var once sync.Once
+//go:embed cdk8s-1.0.0-beta.10.tgz
+var tarball []byte
+var once    sync.Once
 
 // Initialize performs the necessary work for the enclosing
 // module to be loaded in the jsii kernel.
@@ -17,6 +21,6 @@ func Initialize() {
 		constructs.Initialize()
 
 		// Load this library into the kernel
-		rt.Load("cdk8s", "1.0.0-beta.9", tarball)
+		_jsii_.Load("cdk8s", "1.0.0-beta.10", tarball)
 	})
 }
